fix re-rendering problem on the client side
There's a strange behaviour in react-router (v2 / v3):
when the component associated with the route is a promise,
it starts by rendering an empty component and once the
promise is fullfilled, the full components redraws itself.

cf. Router.render() implementation, especially:

    render() {
        [...]
        location = state.location // ← location is null in getInitialState
        [...]
        if (location == null) return null; // Async match
        [...]
    }

To circumvent this, call route.match which will fetch
the component and then only call Rourer.render().