[calendar] use luxon to handle timezones
The database returns timestamps that are later transformed
to JSON then loaded as Date objects in the browser. The Date
objects themselves are correct (i.e. the timestamp is valid)
but if the register date is "1681-04-14T00:00" (UTC), it will be
displayed as "1681-04-13" in the "America/Los Angeles" timezone.

To fix this, encode date as raw objects (parsable with luxon)
with `year`, `month` and `day` properties and use this to format
date in the browser. When we need more complex computations, use
luxon to parse back the date. In that case, Luxon defaults to
"local" timezone which is fine.